implement master leader detection
master.IsLeader() queries the '/master/redirect' endpoint of the
passed-in host to determine if it's the leader, or if another machine in
the cluster is the current leader. This will be used to ensure that
snapd can run on each of the masters, but only collect metrics for the
leader.

Note that you can also check the value of 'master/elected' in the
'/metrics/snapshot' endpoint on the master to determine if that master
is elected the leader. While this would make for a simpler
implementation, it would generate unnecessary load on the master,
building the entirety of the '/metrics/snapshot' JSON document and only
looking for a single value.