refactor test script
Prior to this commit, the test script 'test.sh' was a very procedural
bash script without a lot of structure. This made it a bit confusing to
read through, and resulted in a developer needing to comment out several
lines to disable functionality.

This commit refactors test.sh into several functions, which are then
called based on which test suite is running (unit or integration). This
should make it easier to understand and maintain, while making it easy
to enable / disable features in this script going forward.