split mesos metrics containing "/"
Prior to this commit, metrics being discovered on Mesos masters and
agents via GetMetricTypes() returned the metric key as a string, which
contains forward slashes ("/"). While `snapctl metric list` displayed
these properly, the fact remains that metric keys with "/" are not
allowed in snap and therefore CollectMetrics() was never called.

This commit splits the metric key and appends it to the namespace, which
allows the snap metric catalog to be properly populated.