abstract http client for consistent use
Prior to this commit, we performed the following steps in several
places, largely repeating ourselves:

  * Create a new HTTP client
  * Fetch data from a URL
  * Unmarshal the JSON
  * Return the data or an error

This commit abstracts the HTTP client into a package of its own, and
modifies the master and agent packages to use this instead. This keeps
with the DRY principle and ensures that each API query is consistent.
This is also implemented in a way that the caller simply passes in a
pointer to a variable, so that JSON can still be worked with as if the
calling function was unmarshaling the JSON itself.