(gh-11) add golang-compiled mesos 0.28.1 protobuf
Snap's metric catalog is only populated once: when the plugin is loaded.
Therefore, due to the highly dynamic nature of executors running on a
Mesos agent (they're always changing during normal operation) it isn't
possible to populate the metric catalog and collect metrics by walking
the flat JSON object like we can with the "/metrics/snapshot" endpoint.
I explained this a bit on GitHub (see #11).

In this commit, I've brought in the compiled `mesos.proto` protocol
buffers from Mesos 0.28.1. When using github.com/golang/protobuf and
protoc, these messages are compiled to native Go structs, which allow us
to avoid recreating (and maintaining!) all of these fields ourselves.
I've also added a new task in the Makefile that should allow maintainers
to update this easily in the future.