add source tag with mesos master/agent host:port
When this plugin was updated to use the new snap metric schema, the
source of the metric (the host/IP and port) was no longer available for
collected metrics. The new schema automatically adds a
"plugin_running_on" tag with the value of os.Hostname(), but this could
be inaccurate for two reasons:

  * The Mesos master / agent could be running on a different interface
    than the one returned by os.Hostname()
  * Multiple agents could be running on a single machine, with unique
    ports

This commit adds a new tag "source" with the value of the "master" or
"agent" configuration option.