replace one-off reflect with FromCompositeObject()
Prior to this commit, I implemented a one-off recursive reflection
function to build snap metric namespaces from the ResourceStatistics
struct in the Mesos protobuf. @marcin-krolik mentioned that similar
functionality already exists in snap-plugin-utilities.

This commit replaces the custom / one-off reflect usage with the
implementation in github.com/intelsdi-x/snap-plugin-utilities/ns,
which will be far more maintainable in the future.