make it easy to develop the grafana example
Prior to this commit, it was non-trivial to further develop the Grafana
dashboard that I recently introduced. A developer would need to set up a
Mesos cluster, along with InfluxDB and Grafana, and manually import the
dashboard if he or she wanted to make changes.

This commit installs and configures InfluxDB and Grafana within the
Vagrant development environment so that anyone can easily improve the
dashboard (because it needs it!)