(gh-19) expose metric types for enabled features
Prior to this commit, this plugin exposed all possible metrics by
recursively walking the ResourceStatistics message in mesos_pb2. This
had the side effect of making *all* possible metric types available to
the user via `snapctl metric list`. In reality, the user might find
him/herself in an impossible situation, thinking that a metric type
exists when in-fact it does not.

This commit checks the enabled isolators on a Mesos agent, and deletes
any metrics that don't actually exist.

Note: there's some discussion in
https://github.com/intelsdi-x/snap/issues/936 that using configuration
items in GetMetricTypes() is a bad practice. There's still some
outstanding discussion on that issue, so I'm implementing this for now
until we have a clear path forward. Once consensus is reached in the
Snap community, then we can revisit this particular patch /
functionality.